Fix 4 frontend bugs: remove dead openLeveragedStrategyDirect, liquidate, emergencyWithdraw; fix uint64→uint128 in rebalance
BUG-1: useComposer.openLeveragedStrategyDirect called non-existent contract function — removed
BUG-2: useLendingActions.liquidate called Pool.liquidate which doesn't exist (only liquidateWithProof) — removed
BUG-3: useLendingActions.emergencyWithdraw targeted Pool (only Vault has it) — removed
BUG-4: useRebalance encrypted repay/newBorrow as uint64 instead of uint128 — fixed to uint128

Also updated lending-actions-demo.tsx and use-lending-actions.spec.ts to match.

Constraint: Frontend refactor deferred until backend confirmed ready
Tested: npx tsc --noEmit — exit 0, clean
Not-tested: Runtime tx execution (requires wallet + arb-sepolia)

+++ ui/hooks/use-rebalance.ts
@@ -52,12 +52,12 @@ export function useRebalance() {
     if (!addCollateralEnc[0]) throw new Error("CoFHE returned empty handle for addCollateralAmount");
 
     const repayEnc = (await cofheClient
-      .encryptInputs([Encryptable.uint64(params.repayAmount)])
+      .encryptInputs([Encryptable.uint128(params.repayAmount)])
       .execute()) as InEuint128[];
     if (!repayEnc[0]) throw new Error("CoFHE returned empty handle for repayAmount");
 
     const newBorrowEnc = (await cofheClient
-      .encryptInputs([Encryptable.uint64(params.newBorrowAmount)])
+      .encryptInputs([Encryptable.uint128(params.newBorrowAmount)])
       .execute()) as InEuint128[];
     if (!newBorrowEnc[0]) throw new Error("CoFHE returned empty handle for newBorrowAmount");
